Checkpointing a bunch of changes.
This updates a lot for the category project.  This also
updates a lot of the administrative tools.

Revised the about and club pages, moving much of the
verbiage about ongoing work into a new site status page.
The comment page has been moved into cgi-bin land as it
was getting pretty complicated.